Implement LinkedModelGroupLayout
This implements a new layout that allows adding and removing models and
contains a search bar. Adding and removing will be needed for spawning knobs on
drag-drop (see #5095).

This requires changes in LinkedModelGroup classes, though these classes
will be removed soon (and placed into Lv2/Spa/... directly):

Previously, the models for each processor have been stored in a vector,
now they are in a map (LinkedModelGroups.h). In this step, the
m_linkEnabled also have been pulled from a vector into a map (the same
map now as the models).

Note: If VST does not compile with this commit anymore, simply use
`cmake -DPLUGIN_LIST=x;y;z` to not enable VST. This is an issue that
has been fixed on master already.