Fixes #5194: Fix knobs moving too fast
For systems where QCursor::setPos() has no effect, this fix adds a
workaround for the situation where a knob value is changed via mouse
click and move. The fix simulates that the cursor is moved back to the
original click position.

Before this commit, the cursor was always reset over the knob. Now, it
runs over the screen and tries to flip on the opposite screen side if it
hits the screen border.

This changes the meaning of `Knob::m_origMousePos` from "always being
the position where the user clicked" to "the last position in
mouseMoveEvent". And it renames it to `m_lastMousePos`.

On Apple, `setPos` and cursor hiding are now completely disabled.