Adds midi channel forwarding feature.
Now it's possible to forward received midi events with their original channel, either to another track or to the instrument plugin itself. To do that, user must select channel "--" on the midi output widget. When that channel is selected, all midi events will be forwarded with their original channel and the midi events produced by the track itself will be sent with the default channel 1.