Revert "Adds the new solo behavior as a new LED button"
This reverts commit fdbc8b2712d85cb7cafaf0298531557135cf9968.

After consulting fellow users and devs about this change to the PR, it was decided that adding a third button for the new solo behavior was not the best approach. This reverts the commit so we go back to just changing the solo behavior. Later an option can be added on LMMS settings to choose between the old and new behaviors.