First commit
This branch will be used to implement a new feature to LMMS, where MIDI CC messages support will be added through the use of a rack holding knobs representing each MIDI controller. The changes to values of those knobs will trigger MIDI CC messages to the corresponding track. There will be the possibility of automating the values.

Summary of changes:
	- Creates a new icon to be used on the MIDI CC rack window.
	- Adds the MIDI CC rack window as a new class.
	- Very basic window toggling functionality.
	- Updates CMakeList to compile this new .cpp file.