Adds the enable/disable functionality for MIDI CC
Now the GroupBox LED button will enable or disable the MIDI CC events triggering for the selected InstrumentTrack. The LED button is connected to the selected InstrumentTrack's BoolModel. If it's disabled, the callback function that processes the MIDI CC messages will be called but will return before processing the MIDI CC event. The connection isn't destroyed in that case (probably the overhead of having to recreate the connection every time we enable the MIDI CC for the track doesn't compensate the unnecessary function call). When the LED button is enabled, the MIDI CC event is processed normally.

One behavior that needs to be made clear is the following:

If a value is selected while MIDI CC is enabled (i.e.: 120), then it's disabled, and another value is selected while it's disabled (i.e.: 40), when the MIDI CC is enabled again an event WON'T be triggered until a change in the value is made AFTER it was enabled again. That sounds reasonable and what would be expected from a hardware MIDI controller (when a controller is turned on with a knob on a value, it will only trigger an event for that knob when it's moved to a different value). So I believe it should be kept that way.