Fix SEGFAULT bug and hide MIDI CC rack at start up
A Segfault error was being caused when clicking on the GroupBox label, because instead of setting the model in the GroupBox widget I was setting the model on the LED button widget. The appropriate changes were made to fix the issue.

Also, the MIDI CC rack window is now hidden at start up by default.