Save/Load MIDI controllers models on project file
With this commit, the models related to MIDI CC are saved in the Instrument Track DOM structure. The models are m_midiCCEnable and m_midiCCModel (for each controller). m_midiCCEnable is saved as an attribute on <instrumenttrack> called "enablecc" and m_midiCCModel is saved in a child element called <midicontrollers>.

This allows the user to save projects with CC automation tracks.

TODO:
	- Check if changes on the DataFile class are necessary to make older projects compatible.