Processes MIDI CC rack events as hardware ones
Instead of using processOutEvent to process the MIDI CC events generated by the MIDI CC rack (which is analogous to having them being processed as they were generated by the instrument track) we now use processInEvent. That is similar to simulating a physical MIDI controller because the events will be processed exactly like they were received through the MIDI input ports of the instrument track from a controller.

I decided for this design to avoid having 2 different behaviors: One for physical MIDI controllers and one for the MIDI CC rack. The only possible downside is that MIDI CC events can't be forwarded from one track to another, which is true for physical controller events as well, so it shouldn't be an issue (if in the future we allow forwarding those events we should allow it for both hardware controllers and MIDI CC rack, changing the processInEvent method). The closest we can get from mimicking the behavior of physical controllers the more predictable the feature will be for an user accustomed with the hardware controllers. Any changes to the way LMMS acts when receiving CC events should affect both hardware and MIDI CC rack controllers.