Implements the MidiCCRackView destructor
	I wasn't properly destroying each MidiCCRackView when a track
was removed, causing them to build up. This commit changes the
m_midiCCRackView variable on the InstrumentTrackView class to be a
std::unique_ptr (destroyed when out of scope) and added the appropriate
code on the MidiCCRackView class destructor to remove the parent widget.
	destroyRack() method was removed (it wasn't doing its task
appropriately before and is now unnecessary after the destructor was
fixed) and also unsetModels() which wasn't necessary anymore.