First commit
	This attempts to fix a bug uncovered by #5581, where the wrong
MidiEvent constructor is selected due to ambiguity in the arguments that
are used in the call.
	Part of the solution is to use a enum class on the last
parameter instead of a boolean to avoid implicit conversions and a more
strict method selection through the ADL. One of the parameters type on a
constructor was also changed from int to std::size_t which more
accurately represents it.
	"ignoreOnExport" was replaced with "source", which for now can
be Source::Internal or Source::External, and its value is still used on
the processInEvent to define whether a MidiEvent should be ignored
during the song export.