Fixes bug with pasting of TCOs (#5840)
	TimePos::quantize works for negative values, but ends
up snapping the TCO to the opposite direction. This is because the
snapping happens in the direction of the origin, which is left for
positive values and right for negative values.
	That wasn't accounted for in the pasteSelection method
and we ended up with wrong positions when pasting before the
TCO(s) we copied. This PR attempts a fix by doing the
following:

	- Changing TimePos::quantize to have a boolean parameter that
will define whether it snaps the position up if we are halfway through
the interval or not (this removes the need to use that trick of
subtracting "TimePos::ticksPerBar() * snapSize / 2" from the offset to
try to "bias" it.
	- Checking if we need quantizing at all by checking if the
offset falls outsize the snap grid.
	- If we do need quantizing and the offset is negative we
subtract one snap, since the quantization will move the TCO to the
opposite end (to the right). Then we call quantize.