Use the new SPA v0.0.3 ringbuffers
The new ringbuffers do not allow writing while reading. To prevent
concurrent writing while the plugin is running, midi events are now
cached inside a separate ringbuffer. The implementation is copied from
the Lv2Proc code, and so the code for handling MIDI events was moved
from SpaInstrument to SpaProc.