Fix the buffer being drawn after width
	- Fixes a small bug where the buffer was being drawn past the
end of the widget. That's because when dividing the number of frames by
the width, when it wasn't perfectly divisible the remaining frames where
still being considered for a pixel after the width. That required the
mysterious "+ 2" after the totalPoints to be used. Now those points are
ignored.
	- Adds doxygen comment before method.
	- Adds TODO comment regarding a commented statement using a
parameter that apparently is not used anymore.
	- Renames curFrame to curPixel.
	- Uses curPixel itself for the X calculation.