Bugfixes on sample drawing
	- Fixes a bug where sometimes a higher amount of points was
allocated (because nbFrames/fpp would sometimes be higher than the
width) and those not-initialized extra points would result in a small
glitch in the sample drawing.
	- Fixes a bug where the the last frame wasn't being accounted on
the peak/trough calculation because "frame + i < last" was assuming last
was non inclusive, when it is inclusive ("frame + i <= last").
	- Reverses the change to the x calculation, so it's calculated
straight from curPixel if there are "width" number of points or
proportionally when it's not.