Adds Remove Note action (#4)
* Changes mousePressEvent variables to camelCase

	Changes variable names from PianoRoll::mousePressEvent to
camelCase and changes 2 comments.

* Style fixes on mouseDoubleClickEvent

* Removes magic number and costy loop

	Removes a loop that removed every note except one and replaced
it with a clear() call plus appending the note.
	Removes a magic number used for finding the closest note.

* Small fixes + Code style

	Fixes small bug with the vertical selection mode scrolling.
	Improves logic on the cursor selection of ActionNone.
	Adds update() + return statements for the actions that didn't
have them, and removed the update() from outside the switch.
	Fixes the code style.

* Adds Remove Note action

	Adds a action to remove notes on the PianoRoll, and remove an
unnecessary member variable that was only being used to handle that
action (m_mouseDownRight).
	Also removes unnecessary handling of the play key action on the
ActionNone case switch from the mouseMoveEvent. Adds a update() call on
the ActionPlayKeys (so it updates the cursor) and a break; statement
for consistency. Also fixes the calculation of the velocity.