Don't draw BB editor inside Song Editor
Currently, a small BB editor is drawn inside the Song Editor at high zoom levels. As discussed in #3060 this is unintuitive and appears broken (I've seen several other reports of this as a bug). This PR removes this behavior.

A quick Discord poll implies that 4 other developers agree with this change, but I'll leave this open a bit in case anyone wants to oppose.