Add mix up and down functions
If the current state of SampleBufferV2 was given an audio file
that has more than 2 channels, it would only grab the first 2
samples from each frame. This is an erroneous mixdown to stereo,
as the other samples in the frame are not considered.

To fix this, SampleBufferV2 now has mix up and down to stereo functions,
which are meant to properly handle audio files with more than 2
channels, and simplify the mixing up and down process where the code
is more understandable and readable.

The mixing up to stereo from mono functionality was present already,
but was made more readable with the addition of mixMonoToStereo, rather
than a ternary operator.